plottersim/requirements.txt
Cython
kivy
PyDispatcher
pyserial
